[simulation] refactoring pkg -> utils.go added
